move tryout component
